% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{check_sample_CNA}
\alias{check_sample_CNA}
\title{copy number quality check}
\usage{
check_sample_CNA(
  data,
  outputDir,
  SNV_file,
  LOH,
  tcn_normal_range = c(1.75, 2.3),
  pval = 0.05,
  preprocessing_seed = 123,
  kmeans_nstart = 50
)
}
\description{
Check if a copy-neutral segment is a LOH event by checking the distribution of germline heterozygous mutations
}
