% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{check_sample_LOH}
\alias{check_sample_LOH}
\title{Check if a copy-neutral segment is a LOH event by checking the distribution of germline heterozygous mutations}
\usage{
check_sample_LOH(
  data,
  outputDir,
  SNV_file,
  LOH,
  tcn_normal_range = c(1.8, 2.2),
  pval = 0.05
)
}
\description{
Check if a copy-neutral segment is a LOH event by checking the distribution of germline heterozygous mutations
}
