% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-presence.R
\name{combineMutationCNASamplePresence}
\alias{combineMutationCNASamplePresence}
\title{combine mutation and cna sample presence}
\usage{
combineMutationCNASamplePresence(input_data)
}
\description{
combine mutation and cna sample presence
}
