% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-mutation.R
\name{estimate_multiplicity}
\alias{estimate_multiplicity}
\title{estimate multiplicity for one gene such that all mcf <= 1}
\usage{
estimate_multiplicity(y, n, cncf, cna, threshold = 1)
}
\description{
estimate multiplicity for one gene such that all mcf <= 1
}
