% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-kmeans-clusters.R
\name{estimateClusterMCF}
\alias{estimateClusterMCF}
\title{estimate cluster mcf/cncf by taking the mean}
\usage{
estimateClusterMCF(total_est, clusters)
}
\description{
estimate cluster mcf/cncf by taking the mean
}
