% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-cna.R
\name{estimateCNA1}
\alias{estimateCNA1}
\title{cluster CNAs and resolve CNA state in input_data$y format}
\usage{
estimateCNA1(input_data)
}
\description{
cluster CNAs and resolve CNA state in input_data$y format
}
