% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-cna.R
\name{estimateCNCF1}
\alias{estimateCNCF1}
\title{estimation of copy number cellular fraction in input_data$y format}
\usage{
estimateCNCF1(input_data, cna_est)
}
\description{
estimation of copy number cellular fraction in input_data$y format
}
