% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-mutation.R
\name{estimateMultiplicity}
\alias{estimateMultiplicity}
\title{estimate multiplicity for all mutations}
\usage{
estimateMultiplicity(input_data, cna_est, cncf_est)
}
\description{
estimate multiplicity for all mutations
}
