% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-mutation.R
\name{estimateMultiplicity1}
\alias{estimateMultiplicity1}
\title{estimate mutation cellular fraction
export}
\usage{
estimateMultiplicity1(input_data, cna_est, cncf_est)
}
\description{
estimate multiplicity for all mutations
}
