% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-mutation.R
\name{estimateMutation}
\alias{estimateMutation}
\title{estimate mutation cellular fraction}
\usage{
estimateMutation(input_data, sep_list, cna_est, cncf_est)
}
\description{
estimate mutation cellular fraction
}
