% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans-clustering.R
\name{KmeansClustering}
\alias{KmeansClustering}
\title{run kmeans clusterig to cluster mutations/CNAs with similar mcf/cncf}
\usage{
KmeansClustering(
  total_est,
  total_sep_list,
  Kmax = 5,
  cluster_mutation_thresh = 1,
  cluster_diff_thresh = 0.05
)
}
\description{
run kmeans clusterig to cluster mutations/CNAs with similar mcf/cncf
}
