% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC-main.R
\name{mcmcMain}
\alias{mcmcMain}
\title{run MCMC with added subclonal copy number}
\usage{
mcmcMain(
  mutation_file,
  copy_number_file,
  outputDir,
  SNP_file = NULL,
  stat_file = NULL,
  cytoband_file = NULL,
  pval = 0.05,
  sim_iter = 100,
  max_K = 3,
  min_mutation_per_cluster = 5,
  iterations = 5,
  min_mutation_per_box = 10,
  n.iter = 5000,
  n.burn = 1000,
  thin = 10,
  mc.cores = 8,
  model_type = "spike_and_slab",
  beta.prior = FALSE,
  drop_zero = TRUE,
  inits = list(.RNG.name = "base::Wichmann-Hill", .RNG.seed = 123),
  cluster_diff_thresh = 0.05,
  alt_reads_thresh = 0,
  vaf_thresh = 0,
  cnv_max_dist = 2000,
  cnv_max_percent = 0.3,
  tcn_normal_range = c(1.8, 2.2),
  smooth_cnv = T,
  autosome = T
)
}
\description{
run MCMC with added subclonal copy number
}
