% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC-clustering.R
\name{runMCMCForAllBoxes}
\alias{runMCMCForAllBoxes}
\title{run MCMC for all mutations by sample presence}
\usage{
runMCMCForAllBoxes(sep_list)
}
\description{
run MCMC for all mutations by sample presence
}
