% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-presence.R
\name{separateMutationsByMCF}
\alias{separateMutationsByMCF}
\title{separate mutations by sample presence using mcf}
\usage{
separateMutationsByMCF(mcf)
}
\description{
separate mutations by sample presence using mcf
}
