% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{importFiles}
\alias{importFiles}
\title{read input data file and store in required format}
\usage{
importFiles(
  mutation_file,
  copy_number_file = NULL,
  SNV_file = NULL,
  outputDir = NULL,
  LOH = FALSE,
  purity_min = 0.2,
  alt_reads_thresh = 0,
  vaf_thresh = 0,
  cnv_max_dist = 1e+06,
  cnv_max_percent = 0.3,
  tcn_normal_range = c(1.75, 2.3),
  filter_cnv = T,
  smooth_cnv = F,
  autosome = T,
  pval = 0.05
)
}
\arguments{
\item{mutation_file}{mutation data file; see inst/extdata/examples/*_snv.csv for examples}

\item{copy_number_file}{copy number file; see inst/extdata/examples/*_cna.csv for examples}

\item{SNV_file}{SNV file for germline heterozygous SNVs; see inst/extdata/examples/*_SNP.csv for examples}

\item{outputDir}{output directory for saving output data}
}
\description{
read input data file and store in required format
}
