% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pictographPlus.R
\name{runPICTographPlus}
\alias{runPICTographPlus}
\title{Tumor evolution drives intra-tumor heterogeneity, yet functional clone-specific profiles remain poorly understood. 
We developed PICTographPlus, a computational tool that integrates bulk DNA and RNA sequencing to reconstruct clone-specific transcriptomic profiles and uncover functional transitions between clones, offering a powerful framework for cancer evolution studies.}
\usage{
runPICTographPlus(
  mutation_file,
  rna_file,
  outputDir = NULL,
  copy_number_file = NULL,
  SNV_file = NULL,
  lambda = 0.2,
  GSEA = TRUE,
  GSEA_file = NULL,
  top_K = 5,
  n_permutations = 1000,
  sample_presence = TRUE,
  score = "silhouette",
  max_K = 10,
  min_mutation_per_cluster = 5,
  min_cluster_thresh = 0.05,
  cluster_diff_thresh = 0.05,
  n.iter = 5000,
  n.burn = 1000,
  thin = 10,
  mc.cores = 8,
  inits = list(.RNG.name = "base::Wichmann-Hill", .RNG.seed = 123),
  driverFile = NULL,
  cytobandFile = NULL,
  threshes = NULL,
  LOH = FALSE,
  purity_min = 0.2,
  dual_model = TRUE,
  ploidy = 2,
  pval = 0.05,
  alt_reads_thresh = 0,
  vaf_thresh = 0,
  cnv_max_dist = 1e+06,
  cnv_max_percent = 0.3,
  tcn_normal_range = c(1.75, 2.3),
  filter_cnv = T,
  smooth_cnv = F,
  autosome = T
)
}
\arguments{
\item{mutation_file}{a csv file that include information for SSMs.}

\item{rna_file}{bulk RNA file in integer read counts; rows are samples and columns are genes}

\item{outputDir}{output directory for saving all files.}

\item{copy_number_file}{a csv file that include information for CNA.}

\item{SNV_file}{a csv file that include information for germline heterozygous SNVs.}

\item{lambda}{weights used in deconvolution step; default is 0.2}

\item{GSEA}{whether to perform GSEA analysis; default is TRUE}

\item{GSEA_file}{geneset file in MSigDB .gmt format; the geneset name will show up in plotting}

\item{sample_presence}{whether to use sample presence to separate the mutations. Not applicable if dual_model is set to FALSE and a copy number file is provided.}

\item{score}{scoring function to estimate the number of clusters. silhouette or BIC.}

\item{max_K}{user defined maximum number of clusters.}

\item{min_mutation_per_cluster}{minumum number of mutations in each cluster.}

\item{cluster_diff_thresh}{threshold to merge two clusters.}

\item{n.iter}{number of iterations by JAGS.}

\item{n.burn}{number of burns by JAGS.}

\item{thin}{number of thin by JAGS.}

\item{mc.cores}{number of cores to use for parallel computing; not applicable to windows.}

\item{inits}{additional parameters by JAGS.}
}
\description{
Tumor evolution drives intra-tumor heterogeneity, yet functional clone-specific profiles remain poorly understood. 
We developed PICTographPlus, a computational tool that integrates bulk DNA and RNA sequencing to reconstruct clone-specific transcriptomic profiles and uncover functional transitions between clones, offering a powerful framework for cancer evolution studies.
}
