% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-presence.R
\name{separateMutationsByMCF}
\alias{separateMutationsByMCF}
\title{}
\usage{
separateMutationsByMCF(mcf)
}
\description{
mutationTypes = separateMutationsBySamplePresence(input_data)
  # print(mutationTypes)
}
\details{

}
